BigDecimal для точных расчётов, ErrorResponse, оптимизация БД

+++ commerce/warehouse/src/main/java/ru/yandex/practicum/warehous/dto/BookedProductsDto.java
@@ -3,10 +3,12 @@ package ru.yandex.practicum.warehous.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class BookedProductsDto {
-    private double deliveryWeight;
-    private double deliveryVolume;
+    private BigDecimal deliveryWeight;
+    private BigDecimal deliveryVolume;
     private boolean fragile;
 }
