+++ commerce/warehouse/src/main/java/ru/yandex/practicum/warehous/dto/NewProductInWarehouseRequest.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.warehous.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -9,5 +10,5 @@ public class NewProductInWarehouseRequest {
     private UUID productId;
     private boolean fragile;
     private DimensionDto dimension;
-    private double weight;
+    private BigDecimal weight;
 }
